minor updates to prediction learner

+++ src/Prediction.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 public class Prediction {
 	Book predictedBook;
 	Genre genre;
-	int score;
+	int score; // score will be provided by user and ranges 1 - 5 (5 is the best result)
 	User user;
-	ArrayList<Integer> weightOfCalc;
+	ArrayList<Double> weightOfCalc;
 }
